requirements.txt
numpy
pyyaml
scipy
matplotlib
opencv-python
stable-baselines3==1.6.0
tensorboard
protobuf~=3.19.0
